some refinement updates
git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/trunk@731 4eea3252-f0fb-4393-894d-40516dce545b

+++ src/sgpp/grid/generation/BoundaryGridGenerator.cpp
@@ -33,7 +33,7 @@ void BoundaryGridGenerator::refine(RefinementFunctor* func)
 	refine.free_refine(this->storage, func);
 }
 
-int BoundaryGridGenerator::getNumberOfRefinablePoints()
+size_t BoundaryGridGenerator::getNumberOfRefinablePoints()
 {
 	HashRefinementBoundaries refine;
 	return refine.getNumberOfRefinablePoints(this->storage);
